restrict most LpExpression functions to pub(crate), move respective tests from tests/expression.rs and tests/quote.rs to test mod in variables.rs
